Preserve order of reference resolution
