+++ apis/common/v1/resource.go
@@ -77,12 +77,6 @@ type SecretKeySelector struct {
 	Key string `json:"key"`
 }
 
-// ResolvePolicy is a type for resolve policy.
-type ResolvePolicy string
-
-// ResolutionPolicy is a type for resolution policy.
-type ResolutionPolicy string
-
 // Policy represents the Resolve and Resolution policies of Reference instance.
 type Policy struct {
 	// Resolve specifies when this reference should be resolved. The default
@@ -100,23 +94,6 @@ type Policy struct {
 	// +kubebuilder:default=Required
 	Resolution *ResolutionPolicy `json:"resolution,omitempty"`
 }
-
-const (
-	// ResolvePolicyAlways is a resolve option.
-	// When the ResolvePolicy is set to ResolvePolicyAlways the reference will
-	// be tried to resolve for every reconcile loop.
-	ResolvePolicyAlways ResolvePolicy = "Always"
-
-	// ResolutionPolicyRequired is a resolution option.
-	// When the ResolutionPolicy is set to ResolutionPolicyRequired the execution
-	// could not continue even if the reference cannot be resolved.
-	ResolutionPolicyRequired ResolutionPolicy = "Required"
-
-	// ResolutionPolicyOptional is a resolution option.
-	// When the ReferenceResolutionPolicy is set to ReferencePolicyOptional the
-	// execution could continue even if the reference cannot be resolved.
-	ResolutionPolicyOptional ResolutionPolicy = "Optional"
-)
 
 // IsResolutionPolicyOptional checks whether the resolution policy of relevant reference is Optional.
 func (p *Policy) IsResolutionPolicyOptional() bool {
@@ -139,7 +116,7 @@ type Reference struct {
 	// Name of the referenced object.
 	Name string `json:"name"`
 
-	// Policies of the referenced object.
+	// Policies for referencing.
 	// +optional
 	Policy *Policy `json:"policy,omitempty"`
 }
@@ -171,7 +148,7 @@ type Selector struct {
 	// as the selecting object is selected.
 	MatchControllerRef *bool `json:"matchControllerRef,omitempty"`
 
-	// Policies of the referenced object.
+	// Policies for selection.
 	// +optional
 	Policy *Policy `json:"policy,omitempty"`
 }
